Completed v0.3
Created user authentication.  Users can log in and out and, depending on their permissions, add new recipes and authors.  Users only need to be logged in to create recipes, however they need staff acess to be able to create authors.  If a user tries to create a new author without staff access, they will be redirected to an error page.